Added navigation files
I added the framwork for the files that hte navigation menu takes you to. We just have to figure out what the content is and then we can work on those files. I took out the #s because we needed to direct to something. I also took out the header because we don't what content we are doing yet. I changed the title for each page so that the title is relevant to the content of the page.