Changed some colors, made it look nicer...
I changed some colors so it would match the website theme which was #37a7c0. Added a paragraph and made it centered to state our website name. On the login, I again matched some colors to theme, fixed some spelling errors and fixed some grammar. I also fixed the background and made some things bigger for better seeability. Is that even a word?