Added blog and content.
I added content for World of Gaming, although, I think it should be deleted. It was just 1 line that I thought fit the topic. The titles got messed up for some of the pages and I fixed that. I also added the blog page and I am waiting for content from Vighnesh. I probably fixed some spaces as well.